move `AccessBuilder` to separate file
